Setting the grid System
I tend to create my own grid system since I prefer it to the one that Unity provides. Here's how it works:

GridSystem: You create it and set its size with gridSystem = new GridSystem(10, 10, 2f);. This gives you a 10x10 grid. While it could be easily modified to change its size and other properties, I decided to stick with the parameters you provided to ensure the GameObjects are placed perfectly.

GridDebugObject: This is a script attached to every spawned cell GameObject. It allows you to see the information for a specific cell and provides a small visual representation.

GridObject: This script works in conjunction with the GridSystem. It stores the data generated by the GridSystem and allows you to check that data later.

GridPosition: A struct that represents every position in the grid.

GridSystem: This class contains the data and logic to create the entire grid system and to iterate through it.

I then configured the scene to create the grid system accordingly.